[MOVE] inviteCodeDto common으로 이동
